+++ passes/S011/testdata/src/a/outside_package.go
@@ -1,0 +1,19 @@
+package a
+
+import (
+	"a/schema"
+)
+
+func foutside() {
+	_ = schema.Schema{
+		Computed:         true,
+		DiffSuppressFunc: func() {},
+	}
+
+	_ = map[string]*schema.Schema{
+		"name": {
+			Computed:         true,
+			DiffSuppressFunc: func() {},
+		},
+	}
+}